[refactor] #98 : mp구조 세분화
