+++ src/features/mapView/ui/bottomsheets/MapDetailBottomSheet.tsx
@@ -1,9 +1,9 @@
 import { useState } from "react";
-import { CarDetail } from "./TransferDetail";
-import { FixedButton, Path, TransferDetail } from "./DetailBottomSheet";
-import { TransferType } from "../model";
 import { SnapBottomSheet } from "@/shared/ui";
 import { useEventStore } from "@/shared/stores";
+import { TransferType } from "../../model";
+import { FixedButton, Path, TransferDetail } from "../detailContents";
+import { CarDetail } from "../detail";
 
 export const MapDetailBottomSheet = () => {
   const detailEventData = useEventStore(state => state.detailEventData);
