+++ src/features/mapView/ui/detail/TransferDetail.tsx
