+++ src/features/mapView/ui/detail/index.ts
@@ -1,0 +1,2 @@
+export * from "./TransferDetail";
+export * from "./UserCard";
