+++ src/features/mapView/ui/detailContents/Path.tsx
