+++ src/features/mapView/ui/detailContents/SubwayPath.tsx
