+++ src/features/mapView/ui/detailContents/TransferDetail.tsx
