+++ src/features/mapView/ui/detailContents/WalkPath.tsx
