+++ src/features/mapView/ui/detailContents/index.ts
