+++ src/features/mapView/ui/index.ts
@@ -1,7 +1,4 @@
-export * from "./KakaoMapView";
-export * from "./AddMemberBottomSheet";
-export * from "./MapDetailBottomSheet";
-export * from "./SnapBottomSheet";
-export * from "./DetailKakaoMapView";
-export * from "./TooCloseSheet";
-export * from "./LoginModal";
+export * from "./bottomsheets";
+export * from "./common";
+export * from "./map";
+export * from "./common";
