+++ src/features/mapView/ui/map/MeetingMarker.tsx
